Implement POSIX timer signal handling and validation
- Added validation for POSIX timer signal events, ensuring signal numbers are within valid range.
- Implemented signal delivery for different notification types (SIGEV_SIGNAL, SIGEV_THREAD_ID, SIGEV_NONE).
- Refactored timer tick handling to utilize the new signal delivery method.
- Introduced tests for various timer creation scenarios, including rejection of unsupported notification types.
- Enhanced existing tests for futex and wait syscall functionalities to cover new edge cases and error conditions.
- Updated syscall manager to return appropriate error codes for unsupported operations.

+++ Fiberish.Core/VFS/TTY/TtyDiscipline.cs
@@ -1299,12 +1299,12 @@ public class TtyDiscipline
             return true;
 
         if (vmin > 0 && vtime == 0)
-            return available >= Math.Min((int)vmin, requestedCount);
+            return available >= Math.Min(vmin, requestedCount);
 
         if (vmin == 0 && vtime > 0)
             return available > 0 || timedOut;
 
-        return available >= Math.Min((int)vmin, requestedCount) || (timedOut && available > 0);
+        return available >= Math.Min(vmin, requestedCount) || (timedOut && available > 0);
     }
 
     private int GetReadTimeoutMs()
